Refaktoring kood korda

+++ src/main/java/ee/janek24back/persistence/providerservice/image/ProviderServiceImageRepository.java
@@ -4,6 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProviderServiceImageRepository extends JpaRepository<ProviderServiceImage, Long> {
-
+public interface ProviderServiceImageRepository extends JpaRepository<ProviderServiceImage, Integer> {
 }
